refactor to attach listeners in pages component
