+++ src/tests/contexts/ShortcutProvider.test.tsx
@@ -15,8 +15,6 @@ describe('ShortcutProvider', () => {
 
     mockSDK.canvas.setZoomPercentage = jest.fn().mockImplementation().mockReturnValue({ success: true, code: 200 });
 
-    mockSDK.page.select = jest.fn();
-
     window.StudioUISDK = mockSDK;
     it('triggers the sandbox toggle shortcut', async () => {
         const user = userEvent.setup();
@@ -28,11 +26,6 @@ describe('ShortcutProvider', () => {
                     projectConfig={projectConfig}
                     zoom={100}
                     undoStackState={{ canRedo: true, canUndo: true }}
-                    pages={[
-                        { id: '0', number: 0 },
-                        { id: '1', number: 1 },
-                    ]}
-                    activePageId="0"
                 >
                     <h1>This is a test</h1>
                 </ShortcutProvider>
@@ -52,11 +45,6 @@ describe('ShortcutProvider', () => {
                     projectConfig={projectConfig}
                     zoom={100}
                     undoStackState={{ canRedo: true, canUndo: true }}
-                    pages={[
-                        { id: '0', number: 0 },
-                        { id: '1', number: 1 },
-                    ]}
-                    activePageId="0"
                 >
                     <h1>This is a test</h1>
                 </ShortcutProvider>
@@ -75,11 +63,6 @@ describe('ShortcutProvider', () => {
                     projectConfig={projectConfig}
                     zoom={100}
                     undoStackState={{ canRedo: true, canUndo: true }}
-                    pages={[
-                        { id: '0', number: 0 },
-                        { id: '1', number: 1 },
-                    ]}
-                    activePageId="0"
                 >
                     <h1>This is a test</h1>
                 </ShortcutProvider>
@@ -99,11 +82,6 @@ describe('ShortcutProvider', () => {
                     projectConfig={projectConfig}
                     zoom={zoom}
                     undoStackState={{ canRedo: true, canUndo: true }}
-                    pages={[
-                        { id: '0', number: 0 },
-                        { id: '1', number: 1 },
-                    ]}
-                    activePageId="0"
                 >
                     <h1>This is a test</h1>
                 </ShortcutProvider>
@@ -122,11 +100,6 @@ describe('ShortcutProvider', () => {
                     projectConfig={projectConfig}
                     zoom={zoom}
                     undoStackState={{ canRedo: true, canUndo: true }}
-                    pages={[
-                        { id: '0', number: 0 },
-                        { id: '1', number: 1 },
-                    ]}
-                    activePageId="0"
                 >
                     <h1>This is a test</h1>
                 </ShortcutProvider>
@@ -134,30 +107,5 @@ describe('ShortcutProvider', () => {
         );
         fireEvent.keyDown(screen.getByText('This is a test'), { key: '-', ctrlKey: true });
         expect(mockSDK.canvas.setZoomPercentage).toHaveBeenCalledWith(zoom * 0.875);
-    });
-    it('triggers change page on arrow key press hortcut', async () => {
-        const projectConfig = {} as unknown as ProjectConfig;
-        const user = userEvent.setup();
-
-        render(
-            <AppProvider isDocumentLoaded isAnimationPlaying={false}>
-                <ShortcutProvider
-                    projectConfig={projectConfig}
-                    zoom={100}
-                    undoStackState={{ canRedo: true, canUndo: true }}
-                    pages={[
-                        { id: '0', number: 0 },
-                        { id: '1', number: 1 },
-                    ]}
-                    activePageId="0"
-                >
-                    <h1>This is a test</h1>
-                </ShortcutProvider>
-            </AppProvider>,
-        );
-        await user.keyboard('[ArrowRight]');
-        expect(mockSDK.page.select).toHaveBeenCalledWith('1');
-        await user.keyboard('[ArrowLeft]');
-        expect(mockSDK.page.select).toHaveBeenCalledWith('0');
     });
 });
